+++ threedi_modelchecker/tests/test_checks_other.py
@@ -453,12 +453,11 @@ def test_spatial_index_disabled(empty_sqlite_v4):
 )
 def test_potential_breach_start_end(session, x, y, ok):
     # channel geom: LINESTRING (-71.064544 42.28787, -71.0645 42.287)
+    factories.ChannelFactory(id=1)
     factories.PotentialBreachFactory(
-        the_geom=f"SRID=4326;LINESTRING({x} {y}, -71.064544 42.286)"
+        geom=f"SRID=4326;LINESTRING({x} {y}, -71.064544 42.286)", channel_id=1
     )
-    check = PotentialBreachStartEndCheck(
-        models.PotentialBreach.the_geom, min_distance=1.0
-    )
+    check = PotentialBreachStartEndCheck(models.PotentialBreach.geom, min_distance=1.0)
     invalid = check.get_invalid(session)
     if ok:
         assert len(invalid) == 0
@@ -476,15 +475,15 @@ def test_potential_breach_start_end(session, x, y, ok):
 )
 def test_potential_breach_interdistance(session, x, y, ok):
     # channel geom: LINESTRING (-71.064544 42.28787, -71.0645 42.287)
-    ref = factories.PotentialBreachFactory(
-        the_geom="SRID=4326;LINESTRING(-71.06452 42.2874, -71.0646 42.286)"
+    factories.ChannelFactory(id=1)
+    factories.PotentialBreachFactory(
+        geom="SRID=4326;LINESTRING(-71.06452 42.2874, -71.0646 42.286)", channel_id=1
     )
     factories.PotentialBreachFactory(
-        the_geom=f"SRID=4326;LINESTRING({x} {y}, -71.064544 42.286)",
-        channel=ref.channel,
+        geom=f"SRID=4326;LINESTRING({x} {y}, -71.064544 42.286)", channel_id=1
     )
     check = PotentialBreachInterdistanceCheck(
-        models.PotentialBreach.the_geom, min_distance=1.0
+        models.PotentialBreach.geom, min_distance=1.0
     )
     invalid = check.get_invalid(session)
     if ok:
@@ -494,14 +493,17 @@ def test_potential_breach_interdistance(session, x, y, ok):
 
 
 def test_potential_breach_interdistance_other_channel(session):
+    factories.ChannelFactory(id=1)
+    factories.ChannelFactory(id=2)
     factories.PotentialBreachFactory(
-        the_geom="SRID=4326;LINESTRING(-71.06452 42.2874, -71.0646 42.286)"
+        geom="SRID=4326;LINESTRING(-71.06452 42.2874, -71.0646 42.286)", channel_id=1
     )
     factories.PotentialBreachFactory(
-        the_geom="SRID=4326;LINESTRING(-71.06452 42.287401, -71.064544 42.286)"
+        geom="SRID=4326;LINESTRING(-71.06452 42.287401, -71.064544 42.286)",
+        channel_id=2,
     )
     check = PotentialBreachInterdistanceCheck(
-        models.PotentialBreach.the_geom, min_distance=1.0
+        models.PotentialBreach.geom, min_distance=1.0
     )
     invalid = check.get_invalid(session)
     assert len(invalid) == 0